first realistic world physics, with moving ship, world, and stars, along with continuous world scrolling in both directions.

+++ ui.h
@@ -28,8 +28,8 @@ class UI {
   virtual void line(int x1, int y1, int x2, int y2) = 0;
 
   virtual bool readButton(unsigned char s) = 0;
-  virtual int readJoystickX() = 0;
-  virtual int readJoystickY() = 0;
+  virtual float readJoystickX() = 0;
+  virtual float readJoystickY() = 0;
 
   virtual int width() { return _width; }
   virtual int height() { return _height; }
